+++ CHANGELOG.md
@@ -8,4 +8,21 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 * Release for rebranded SDK.
 
+### Remove
+* Remove deprecated methods.
+    * getAmountTotal
+    * setAmountTotal
+    * getCardHolder
+    * setHolder
+    * cancel
+    * cancelAllCharges
+    * cancelAuthorization
+    * getResource
+    * fetchResource
+* Remove deprecated constants.
+    * API_ERROR_AUTHORIZE_ALREADY_CANCELLED
+    * API_ERROR_CHARGE_ALREADY_CHARGED_BACK
+    * API_ERROR_BASKET_ITEM_IMAGE_INVALID_EXTENSION
+    * ENV_VAR_NAME_DISABLE_TEST_LOGGING
+
 [1.1.0.0]: https://github.com/unzerdev/php-sdk/compare/1260b8314af1ac461e33f0cfb382ffcd0e87c105..1.1.0.0